% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{label_A}
\alias{label_A}
\title{Label the A matrix by COSMIC Signatures}
\usage{
label_A(A, P)
}
\arguments{
\item{A}{Estimate of the A matrix.}

\item{P}{Corresponding estimate of the signatures matrix.}
}
\value{
The A matrix with columns labeled as indicated above.
}
\description{
This labels each column of an estimated A matrix by the closest COSMIC v3.2 signature,
as long as the cosine similarity is above 0.7. Note that if the cosine similarity is
below 0.95, "Pseudo" will be appended to the signature name. Duplicates will be indicated
by suffixes.
}
