% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{proportion_A}
\alias{proportion_A}
\title{Annotate signatures by proportion with high exposures.}
\usage{
proportion_A(out, thresh = 0.05, label = T)
}
\arguments{
\item{out}{Output of sampler.}

\item{thresh}{Threshold for relative exposures; default 0.05.}

\item{label}{Whether the A matrix should be annotated by its closest match to
COSMIC v3.2; default TRUE.}
}
\value{
The A matrix, where each entry indicates the proportion of samples in that study
with relative exposures above `thresh`.
}
\description{
This populates each entry of the A matrix by the proportion of samples in that study
with relative exposures above a specified threshold.
}
