% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recovery_Discovery_Covariates.R
\name{recovery_discovery_cov_fit}
\alias{recovery_discovery_cov_fit}
\title{Run the Recovery-Discovery Sampler with Covariates}
\usage{
recovery_discovery_cov_fit(
  M,
  cov,
  lambda_p = 0.5,
  a_p = 5,
  b_p = 0.05,
  lambda_e = 0.1,
  a_e = 5,
  b_e = 0.01,
  a_er = 3,
  b_er = 0.01,
  lambda_er = 0.5
)
}
\arguments{
\item{M}{List of matrices corresponding to each study, in features-by-samples format. The order
of the mutational categories must match that of data('cosmic').}

\item{cov}{List of covariates matrices corresponding to each study,
in covariates-by-samples format; the same covariates in the same order should be supplied for
each study. Covariates must be numeric.}

\item{lambda_p}{Lambda_p, set to 0.5 by default.}

\item{a_p}{a_p, set to 5 by default.}

\item{b_p}{b_p, set to 0.05 by default.}

\item{lambda_e}{lambda_e, set to 0.1 by default.}

\item{a_e}{a_e, set to 5 by default.}

\item{b_e}{b_e, set to 0.01 by default.}

\item{a_er}{a_er, set to 3 by default.}

\item{b_er}{b_er, set to 0.01 by default.}

\item{lambda_er}{lambda_er, set to 0.5 by default.}
}
\value{
A named list describing the best solution:
\itemize{
 \item{BIC: }{the BIC value of the solution.}
 \item{A: }{a list of the point estimates of the A matrices, indicating whether each signature is
 present in each study, corresponding to the discovery and recovery components respectively.}
 \item{P: }{a list of the posterior medians of the signatures matrices, corresponding to the
 discovery and recovery components respectively.}
 \item{E: }{a list of lists of the posterior medians of the exposures matrices for each study s,
 subsetted to just the signatures indicated by the A matrices to belong to that study,
 corresponding to the discovery and recovery components respectively.}
 \item{indics: }{a list of lists of the posterior inclusion probabilities for each study s,
 describing the probability that each covariate affects each signature indicated by A
 to belong to that study, corresponding to the discovery and recovery components respectively.}
 \item{coefs: }{a list of lists of the coefficients, conditional on inclusion, for each study s,
 describing the contribution of each covariate to each signature indicated by A to belong
 to that study, corresponding to the discovery and recovery components respectively.}
 \item{W: }{a list of the normalizing constants for each study; note that the input counts
 matrix M_s for each study s can be modeled as Poisson(W_s * P_s E_s), where P_s and E_s
 contain just the signatures that belong to study s.}
 }
}
\description{
This runs the recovery-discovery sampler for multi-study NMF using COSMIC v3.2 priors,
with covariates. Importantly, the order of mutational categories of the input data has
to match that of the COSMIC reference, see data('cosmic').
}
